move avatar store
