+++ backend/app/store/avatar/testdata/circles.png
